+++ spec/types/ref.ts
@@ -2,7 +2,7 @@ import * as Spec from './spec'
 import { Type } from './typebox'
 
 {
-    const T = Type.String()
+    const T = Type.String({ $id: 'T' })
 
     const R = Type.Ref(T)